// NAFClient/src/types/enum/progress.ts
export enum Progress {
  OPEN = 0, // NO APPROVERS YET
  IN_PROGRESS = 1, // AN APPROVER APPROVED THE REQUEST AND NOW WAITING FOR THE NEXT APPROVER/S
  REJECTED = 2, // AN APPROVER REJECTED THE REQUEST
  FOR_SCREENING = 3, // ALL APPROVERS APPROVED THE REQUEST AND THE TECHNICAL TEAM WILL PROCEED WITH THE PREPARATION
  ACCOMPLISHED = 4, // ALL APPROVERS APPROVED THE REQUEST AND THE TECHNICAL TEAM DELIVERED THE REQUEST
  NOT_ACCOMPLISHED = 5, // THE REQUEST IS REJECTED AND THE REQUESTOR CLOSED THE REQUEST
}